backend/custom_model/food_recognition_api.py
"""
Placeholder for your custom food recognition model API.
This is where you'll integrate your custom-trained model.
"""

from fastapi import FastAPI, UploadFile, File, HTTPException
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
import numpy as np
from PIL import Image
import io
import json
import logging

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI(title="Food Recognition Model API", version="1.0.0")

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# TODO: Load your custom trained model here
# model = load_your_custom_model()

@app.get("/health")
async def health_check():
    return {"status": "healthy", "model": "food_recognition_v1"}

@app.post("/analyze")
async def analyze_food_image(image: UploadFile = File(...)):
    """
    Analyze food image and return nutritional breakdown.
    
    This is a placeholder implementation. Replace with your actual model inference.
    """
    try:
        # Validate image
        if not image.content_type.startswith('image/'):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read and process image
        image_data = await image.read()
        pil_image = Image.open(io.BytesIO(image_data))
        
        # TODO: Replace this with your actual model inference
        # processed_image = preprocess_image(pil_image)
        # predictions = model.predict(processed_image)
        # nutritional_data = post_process_predictions(predictions)
        
        # Mock response for development
        mock_result = {
            "food_items": [
                {
                    "name": "Grilled Salmon",
                    "calories": 206,
                    "protein": 22.0,
                    "carbs": 0.0,
                    "fat": 12.0,
                    "portion_size": "100g",
                    "confidence": 0.92
                },
                {
                    "name": "Steamed Broccoli",
                    "calories": 34,
                    "protein": 2.8,
                    "carbs": 7.0,
                    "fat": 0.4,
                    "portion_size": "100g",
                    "confidence": 0.88
                },
                {
                    "name": "Brown Rice",
                    "calories": 111,
                    "protein": 2.6,
                    "carbs": 23.0,
                    "fat": 0.9,
                    "portion_size": "100g",
                    "confidence": 0.85
                }
            ],
            "total_calories": 351,
            "total_protein": 27.4,
            "total_carbs": 30.0,
            "total_fat": 13.3,
            "analysis_confidence": 0.88,
            "processing_time_ms": 1250
        }
        
        logger.info(f"Analyzed image: {image.filename}")
        return mock_result
        
    except Exception as e:
        logger.error(f"Food analysis error: {str(e)}")
        raise HTTPException(status_code=500, detail=str(e))

def preprocess_image(image: Image.Image) -> np.ndarray:
    """
    Preprocess image for model inference.
    TODO: Implement your preprocessing pipeline.
    """
    # Resize image
    image = image.resize((224, 224))
    
    # Convert to RGB if needed
    if image.mode != 'RGB':
        image = image.convert('RGB')
    
    # Convert to numpy array and normalize
    image_array = np.array(image) / 255.0
    
    # Add batch dimension
    image_array = np.expand_dims(image_array, axis=0)
    
    return image_array

def post_process_predictions(predictions) -> dict:
    """
    Post-process model predictions to extract nutritional information.
    TODO: Implement your post-processing logic.
    """
    # This is where you would convert your model's raw predictions
    # into the structured nutritional data format
    pass

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8001)
