backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.0
google-generativeai==0.3.2
firebase-admin==6.4.0
upstash-vector==0.4.0
aiohttp==3.9.1
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
